+++ zyx-core/src/utils.rs
@@ -66,7 +66,7 @@ pub fn get_dtype(nodes: &[Node], mut x: Id) -> DType {
 }
 
 /// Puts graph of nodes into dot language for visualization
-pub fn plot_graph_dot(ids: &BTreeSet<Id>, nodes: &[Node], rcs: &[u16]) -> alloc::string::String {
+pub fn plot_graph_dot(ids: &BTreeSet<Id>, nodes: &[Node], rcs: &[u32]) -> alloc::string::String {
     //let ids = &(0..nodes.len()).map(crate::tensor::id).collect::<BTreeSet<Id>>();
     use alloc::{format, string::String};
     use core::fmt::Write;
